Split src/main.rs into src/main.rs + src/app/ folder
Mirrors the config split (PR #469). main.rs at 2571 lines stopped
predicting what it contained — an audio resampler, a JSON formatter
for Waybar, a meeting CLI with 11 action arms, and a 290-line CLI
override block all lived alongside fn main. Rule 4 from
docs/REFACTORING.md (file name no longer predicts contents) and
Rule 3 (every new flag forces another override stanza) both held.

main.rs is now a 75-line shim: sigill install, sigpipe reset, log
init, Cli::parse, config load, then app::run. The remainder lives
under src/app/ in 14 files organised by subcommand (record, status,
meeting, info, config_show, config_set_engine, transcribe_file,
updates, macos) plus shared plumbing (daemon_pid, overrides,
dispatch, sigpipe).

Behaviour-preserving: 770 lib tests + 3 bin tests pass, clippy clean,
all 9 subcommand --help outputs byte-identical to baseline captured
from origin/dev. Three orchestrators are new (app::run,
apply_cli_overrides, dispatch::dispatch); they're plumbing, not logic.

The three regression tests at the bottom of main.rs moved to live
next to the code they pin: pid_file_path_matches_send_record_command
in app/daemon_pid.rs, record_cancel_leaves_idle_not_stopped in
app/status.rs, config_displays_meeting_section in app/config_show.rs.

Per the plan note's "not in scope": no dedup of the engine-string
parsing (duplicated in overrides + Transcribe arm), the 7 paired
enable/disable override blocks, format_state_json's string concat,
or the daemon-liveness helpers. Those are deferred to a follow-up
commit on this branch.
