+++ src/main.rs
@@ -3,45 +3,18 @@
 //! Run with `voxtype` or `voxtype daemon` to start the daemon.
 //! Use `voxtype setup` to check dependencies and download models.
 //! Use `voxtype transcribe <file>` to transcribe an audio file.
+//!
+//! The binary entry point is intentionally thin: install the SIGILL handler
+//! before any other code, reset SIGPIPE, parse CLI, set up logging, load
+//! config, then hand off to `app::run`. Every long handler (status, meeting,
+//! record, …) lives under `src/app/`.
 
+mod app;
+
+use app::sigpipe;
 use clap::Parser;
-use std::path::PathBuf;
 use tracing_subscriber::EnvFilter;
-#[cfg(target_os = "macos")]
-use voxtype::menubar;
-use voxtype::{
-    config, config_set, cpu, daemon, meeting, setup, transcribe, vad, Cli, Commands, ConfigAction,
-    ConfigSetKey, InfoAction, MeetingAction, RecordAction, SetupAction,
-};
-
-/// Parse a comma-separated list of driver names into OutputDriver vec
-fn parse_driver_order(s: &str) -> Result<Vec<config::OutputDriver>, String> {
-    s.split(',')
-        .map(|d| d.trim().parse::<config::OutputDriver>())
-        .collect()
-}
-
-/// Check if running as root and warn for commands that don't need elevated privileges.
-/// Returns true if running as root.
-fn warn_if_root(command_name: &str) -> bool {
-    // SAFETY: getuid() is always safe to call
-    let is_root = unsafe { libc::getuid() } == 0;
-    if is_root {
-        eprintln!(
-            "Warning: Running 'voxtype setup {}' as root is not recommended.",
-            command_name
-        );
-        eprintln!("  - Models will download to /root/.local/share/voxtype/ instead of your user directory");
-        eprintln!(
-            "  - Config changes will apply to /root/.config/voxtype/ instead of your user config"
-        );
-        eprintln!("  - Cannot restart your user's voxtype daemon from root");
-        eprintln!();
-        eprintln!("Run without sudo: voxtype setup {}", command_name);
-        eprintln!();
-    }
-    is_root
-}
+use voxtype::{config, cpu, Cli, Commands};
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
@@ -51,7 +24,7 @@ async fn main() -> anyhow::Result<()> {
 
     // Reset SIGPIPE to default behavior (terminate silently) to avoid panics
     // when output is piped through commands like `head` that close the pipe early
-    reset_sigpipe();
+    sigpipe::reset_sigpipe();
 
     let cli = Cli::parse();
 
@@ -96,2476 +69,7 @@ async fn main() -> anyhow::Result<()> {
         .clone()
         .or_else(config::Config::resolve_existing_path)
         .or_else(config::Config::default_path);
-    let mut config = config::load_config(cli.config.as_deref())?;
+    let config = config::load_config(cli.config.as_deref())?;
 
-    // Apply CLI overrides
-    if cli.clipboard {
-        config.output.mode = config::OutputMode::Clipboard;
-    }
-    if cli.paste {
-        config.output.mode = config::OutputMode::Paste;
-    }
-    if cli.restore_clipboard {
-        config.output.restore_clipboard = true;
-    }
-    if let Some(delay) = cli.restore_clipboard_delay_ms {
-        config.output.restore_clipboard_delay_ms = delay;
-    }
-    let top_level_model = cli.model.clone();
-    if let Some(model) = cli.model {
-        if setup::model::is_valid_model(&model) {
-            config.whisper.model = model;
-        } else {
-            let default_model = &config.whisper.model;
-            tracing::warn!(
-                "Unknown model '{}', using default model '{}'",
-                model,
-                default_model
-            );
-            // Send desktop notification
-            voxtype::notification::send_sync(
-                "Voxtype: Invalid Model",
-                &format!("Unknown model '{}', using '{}'", model, default_model),
-            );
-        }
-    }
-    if let Some(engine) = cli.engine {
-        match engine.to_lowercase().as_str() {
-            "whisper" => config.engine = config::TranscriptionEngine::Whisper,
-            "parakeet" => config.engine = config::TranscriptionEngine::Parakeet,
-            "moonshine" => config.engine = config::TranscriptionEngine::Moonshine,
-            "sensevoice" => config.engine = config::TranscriptionEngine::SenseVoice,
-            "paraformer" => config.engine = config::TranscriptionEngine::Paraformer,
-            "dolphin" => config.engine = config::TranscriptionEngine::Dolphin,
-            "omnilingual" => config.engine = config::TranscriptionEngine::Omnilingual,
-            "cohere" => config.engine = config::TranscriptionEngine::Cohere,
-            "soniox" => config.engine = config::TranscriptionEngine::Soniox,
-            _ => {
-                eprintln!(
-                    "Error: Invalid engine '{}'. Valid options: whisper, parakeet, moonshine, sensevoice, paraformer, dolphin, omnilingual, cohere, soniox",
-                    engine
-                );
-                std::process::exit(1);
-            }
-        }
-    }
-
-    // Hotkey overrides
-    if let Some(hotkey) = cli.hotkey {
-        config.hotkey.key = hotkey;
-    }
-    if cli.toggle {
-        config.hotkey.mode = config::ActivationMode::Toggle;
-    }
-    if cli.no_hotkey {
-        config.hotkey.enabled = false;
-    }
-    if let Some(cancel_key) = cli.cancel_key {
-        config.hotkey.cancel_key = Some(cancel_key);
-    }
-    if let Some(model_modifier) = cli.model_modifier {
-        config.hotkey.model_modifier = Some(model_modifier);
-    }
-
-    // Whisper overrides
-    if let Some(delay) = cli.pre_type_delay {
-        config.output.pre_type_delay_ms = delay;
-    }
-    if let Some(delay) = cli.wtype_delay {
-        tracing::warn!("--wtype-delay is deprecated, use --pre-type-delay instead");
-        config.output.pre_type_delay_ms = delay;
-    }
-    if cli.no_whisper_context_optimization {
-        config.whisper.context_window_optimization = false;
-    }
-    if let Some(prompt) = cli.initial_prompt {
-        config.whisper.initial_prompt = Some(prompt);
-    }
-    if let Some(lang) = cli.language {
-        config.whisper.language = config::LanguageConfig::from_comma_separated(&lang);
-    }
-    if cli.translate {
-        config.whisper.translate = true;
-    }
-    if let Some(threads) = cli.threads {
-        config.whisper.threads = Some(threads);
-    }
-    if cli.gpu_isolation {
-        config.whisper.gpu_isolation = true;
-    }
-    if let Some(gpu_device) = cli.gpu_device {
-        config.whisper.gpu_device = Some(gpu_device);
-    }
-    if cli.flash_attention {
-        config.whisper.flash_attention = true;
-    }
-    if cli.on_demand_loading {
-        config.whisper.on_demand_loading = true;
-    }
-    if let Some(ref mode) = cli.whisper_mode {
-        match mode.to_lowercase().as_str() {
-            "local" => config.whisper.mode = Some(config::WhisperMode::Local),
-            "remote" => config.whisper.mode = Some(config::WhisperMode::Remote),
-            "cli" => config.whisper.mode = Some(config::WhisperMode::Cli),
-            _ => {
-                eprintln!(
-                    "Error: Invalid whisper mode '{}'. Valid options: local, remote, cli",
-                    mode
-                );
-                std::process::exit(1);
-            }
-        }
-    }
-    if let Some(model) = cli.secondary_model {
-        config.whisper.secondary_model = Some(model);
-    }
-    if cli.eager_processing {
-        config.whisper.eager_processing = true;
-    }
-    if let Some(endpoint) = cli.remote_endpoint {
-        config.whisper.remote_endpoint = Some(endpoint);
-    }
-    if let Some(model) = cli.remote_model {
-        config.whisper.remote_model = Some(model);
-    }
-    if let Some(key) = cli.remote_api_key {
-        config.whisper.remote_api_key = Some(key);
-    }
-
-    // Soniox overrides
-    if let Some(key) = cli.soniox_api_key {
-        config
-            .soniox
-            .get_or_insert_with(config::SonioxConfig::default)
-            .api_key = Some(key);
-    }
-
-    // Audio overrides
-    if let Some(device) = cli.audio_device {
-        config.audio.device = device;
-    }
-    if let Some(max_dur) = cli.max_duration {
-        config.audio.max_duration_secs = max_dur;
-    }
-    if cli.audio_feedback {
-        config.audio.feedback.enabled = true;
-    }
-    if cli.no_audio_feedback {
-        config.audio.feedback.enabled = false;
-    }
-    if cli.pause_media {
-        config.audio.pause_media = true;
-    }
-
-    // Output overrides
-    if let Some(append_text) = cli.append_text {
-        config.output.append_text = Some(append_text);
-    }
-    if cli.wtype_shift_prefix {
-        config.output.wtype_shift_prefix = true;
-    }
-    if let Some(ref driver_str) = cli.driver {
-        match parse_driver_order(driver_str) {
-            Ok(drivers) => {
-                config.output.driver_order = Some(drivers);
-            }
-            Err(e) => {
-                eprintln!("Error: {}", e);
-                std::process::exit(1);
-            }
-        }
-    }
-    if cli.auto_submit {
-        config.output.auto_submit = true;
-    }
-    if cli.no_auto_submit {
-        config.output.auto_submit = false;
-    }
-    if cli.shift_enter_newlines {
-        config.output.shift_enter_newlines = true;
-    }
-    if cli.no_shift_enter_newlines {
-        config.output.shift_enter_newlines = false;
-    }
-    if cli.smart_auto_submit {
-        config.text.smart_auto_submit = true;
-    }
-    if cli.no_smart_auto_submit {
-        config.text.smart_auto_submit = false;
-    }
-    if let Some(delay) = cli.type_delay {
-        config.output.type_delay_ms = delay;
-    }
-    if cli.fallback_to_clipboard {
-        config.output.fallback_to_clipboard = true;
-    }
-    if cli.no_fallback_to_clipboard {
-        config.output.fallback_to_clipboard = false;
-    }
-    if cli.spoken_punctuation {
-        config.text.spoken_punctuation = true;
-    }
-    if cli.filter_fillers {
-        config.text.filter_filler_words = true;
-    }
-    if cli.no_filter_fillers {
-        config.text.filter_filler_words = false;
-    }
-    if let Some(keys) = cli.paste_keys {
-        config.output.paste_keys = Some(keys);
-    }
-    if let Some(layout) = cli.dotool_xkb_layout {
-        config.output.dotool_xkb_layout = Some(layout);
-    }
-    if let Some(variant) = cli.dotool_xkb_variant {
-        config.output.dotool_xkb_variant = Some(variant);
-    }
-    if let Some(layout) = cli.eitype_xkb_layout {
-        config.output.eitype_xkb_layout = Some(layout);
-    }
-    if let Some(variant) = cli.eitype_xkb_variant {
-        config.output.eitype_xkb_variant = Some(variant);
-    }
-    if let Some(path) = cli.file_path {
-        config.output.file_path = Some(path);
-    }
-    if let Some(ref mode) = cli.file_mode {
-        match mode.to_lowercase().as_str() {
-            "overwrite" => config.output.file_mode = config::FileMode::Overwrite,
-            "append" => config.output.file_mode = config::FileMode::Append,
-            _ => {
-                eprintln!(
-                    "Error: Invalid file mode '{}'. Valid options: overwrite, append",
-                    mode
-                );
-                std::process::exit(1);
-            }
-        }
-    }
-    if let Some(cmd) = cli.pre_output_command {
-        config.output.pre_output_command = Some(cmd);
-    }
-    if let Some(cmd) = cli.post_output_command {
-        config.output.post_output_command = Some(cmd);
-    }
-    if let Some(cmd) = cli.pre_recording_command {
-        config.output.pre_recording_command = Some(cmd);
-    }
-    if cli.wait_for_modifier_release {
-        config.output.wait_for_modifier_release = true;
-    }
-    if cli.no_wait_for_modifier_release {
-        config.output.wait_for_modifier_release = false;
-    }
-    if let Some(ms) = cli.modifier_release_timeout_ms {
-        config.output.modifier_release_timeout_ms = ms;
-    }
-
-    // VAD overrides
-    if cli.vad {
-        config.vad.enabled = true;
-    }
-    if let Some(threshold) = cli.vad_threshold {
-        config.vad.threshold = threshold.clamp(0.0, 1.0);
-    }
-    if let Some(ref backend) = cli.vad_backend {
-        config.vad.backend = match backend.to_lowercase().as_str() {
-            "auto" => config::VadBackend::Auto,
-            "energy" => config::VadBackend::Energy,
-            "whisper" => config::VadBackend::Whisper,
-            _ => {
-                eprintln!(
-                    "Unknown VAD backend '{}'. Valid options: auto, energy, whisper",
-                    backend
-                );
-                std::process::exit(1);
-            }
-        };
-    }
-    if let Some(min_speech) = cli.vad_min_speech_ms {
-        config.vad.min_speech_duration_ms = min_speech;
-    }
-
-    // On macOS, detect if launched as app bundle executable (no subcommand, binary inside .app)
-    #[cfg(target_os = "macos")]
-    let default_command = if cli.command.is_none() {
-        std::env::current_exe()
-            .ok()
-            .and_then(|p| p.to_str().map(|s| s.contains(".app/Contents/MacOS/")))
-            .unwrap_or(false)
-            .then_some(Commands::AppLaunch)
-            .unwrap_or(Commands::Daemon)
-    } else {
-        Commands::Daemon // unused, cli.command is Some
-    };
-    #[cfg(not(target_os = "macos"))]
-    let default_command = Commands::Daemon;
-
-    // Run the appropriate command
-    match cli.command.unwrap_or(default_command) {
-        Commands::Daemon => {
-            let mut daemon = daemon::Daemon::new(config, config_path);
-            daemon.run().await?;
-        }
-        #[cfg(target_os = "macos")]
-        Commands::Menubar => {
-            let state_file = config
-                .resolve_state_file()
-                .ok_or_else(|| anyhow::anyhow!("state_file not configured"))?;
-            menubar::run(state_file);
-            // Note: menubar::run() never returns (runs macOS event loop)
-        }
-        #[cfg(target_os = "macos")]
-        Commands::AppLaunch => {
-            // Launched by Voxtype.app: start daemon in background, run menubar in foreground.
-            // The binary must be the CFBundleExecutable (not exec'd from a wrapper script)
-            // so macOS Control Center can register the status bar scene correctly.
-            let logs_dir = dirs::home_dir()
-                .map(|h| h.join("Library/Logs/voxtype"))
-                .unwrap_or_else(|| std::path::PathBuf::from("/tmp/voxtype"));
-            let _ = std::fs::create_dir_all(&logs_dir);
-
-            // First-launch auto-setup: create config and download model if needed
-            first_launch_setup(&config).await;
-
-            // Kill any existing instances
-            let _ = std::process::Command::new("pkill")
-                .args(["-9", "-f", "voxtype-bin daemon"])
-                .status();
-            let _ = std::process::Command::new("pkill")
-                .args(["-9", "-f", "voxtype-bin menubar"])
-                .status();
-            let _ = std::fs::remove_file("/tmp/voxtype/voxtype.lock");
-            let _ = std::fs::remove_file("/tmp/voxtype/menubar.lock");
-
-            // Start daemon as a child process with logging
-            let exe = std::env::current_exe()?;
-            let stdout = std::fs::File::create(logs_dir.join("stdout.log"))?;
-            let stderr = std::fs::File::create(logs_dir.join("stderr.log"))?;
-            let _daemon = std::process::Command::new(&exe)
-                .arg("daemon")
-                .stdout(stdout)
-                .stderr(stderr)
-                .spawn()?;
-
-            // Run menubar in this process (keeps the app alive with menu bar icon)
-            let state_file = config
-                .resolve_state_file()
-                .ok_or_else(|| anyhow::anyhow!("state_file not configured"))?;
-            menubar::run(state_file);
-        }
-
-        Commands::Transcribe { file, engine } => {
-            if let Some(engine_name) = engine {
-                match engine_name.to_lowercase().as_str() {
-                    "whisper" => config.engine = config::TranscriptionEngine::Whisper,
-                    "parakeet" => config.engine = config::TranscriptionEngine::Parakeet,
-                    "moonshine" => config.engine = config::TranscriptionEngine::Moonshine,
-                    "sensevoice" => config.engine = config::TranscriptionEngine::SenseVoice,
-                    "paraformer" => config.engine = config::TranscriptionEngine::Paraformer,
-                    "dolphin" => config.engine = config::TranscriptionEngine::Dolphin,
-                    "omnilingual" => config.engine = config::TranscriptionEngine::Omnilingual,
-                    "cohere" => config.engine = config::TranscriptionEngine::Cohere,
-                    "soniox" => config.engine = config::TranscriptionEngine::Soniox,
-                    _ => {
-                        eprintln!("Error: Invalid engine '{}'. Valid options: whisper, parakeet, moonshine, sensevoice, paraformer, dolphin, omnilingual, cohere, soniox", engine_name);
-                        std::process::exit(1);
-                    }
-                }
-            }
-            transcribe_file(&config, &file)?;
-        }
-
-        Commands::TranscribeWorker {
-            model,
-            language,
-            translate,
-            threads,
-        } => {
-            // Internal command: run transcription worker process
-            // This is spawned by the daemon when gpu_isolation is enabled
-            // Use command-line overrides if provided, otherwise use config
-            let mut whisper_config = config.whisper.clone();
-            if let Some(m) = model {
-                whisper_config.model = m;
-            }
-            if let Some(l) = language {
-                // Parse comma-separated language string back to LanguageConfig
-                whisper_config.language = config::LanguageConfig::from_comma_separated(&l);
-            }
-            if translate {
-                whisper_config.translate = true;
-            }
-            if let Some(t) = threads {
-                whisper_config.threads = Some(t);
-            }
-            transcribe::worker::run_worker(&whisper_config)?;
-        }
-
-        Commands::Setup {
-            action,
-            download,
-            model,
-            quiet,
-            no_post_install,
-        } => {
-            match action {
-                Some(SetupAction::Check) => {
-                    warn_if_root("check");
-                    setup::run_checks(&config).await?;
-                }
-                Some(SetupAction::Systemd { uninstall, status }) => {
-                    warn_if_root("systemd");
-                    if status {
-                        setup::systemd::status().await?;
-                    } else if uninstall {
-                        setup::systemd::uninstall().await?;
-                    } else {
-                        setup::systemd::install().await?;
-                    }
-                }
-                #[cfg(target_os = "macos")]
-                Some(SetupAction::Launchd { uninstall, status }) => {
-                    if status {
-                        setup::launchd::status().await?;
-                    } else if uninstall {
-                        setup::launchd::uninstall().await?;
-                    } else {
-                        setup::launchd::install().await?;
-                    }
-                }
-                #[cfg(target_os = "macos")]
-                Some(SetupAction::AppBundle { uninstall, status }) => {
-                    if status {
-                        setup::app_bundle::status().await?;
-                    } else if uninstall {
-                        setup::app_bundle::uninstall().await?;
-                    } else {
-                        setup::app_bundle::install().await?;
-                    }
-                }
-                #[cfg(target_os = "macos")]
-                Some(SetupAction::Hammerspoon {
-                    install,
-                    show,
-                    hotkey,
-                    toggle,
-                }) => {
-                    setup::hammerspoon::run(install, show, &hotkey, toggle).await?;
-                }
-                #[cfg(target_os = "macos")]
-                Some(SetupAction::Macos) => {
-                    setup::macos::run().await?;
-                }
-                Some(SetupAction::Waybar {
-                    json,
-                    css,
-                    install,
-                    uninstall,
-                }) => {
-                    warn_if_root("waybar");
-                    if install {
-                        setup::waybar::install()?;
-                    } else if uninstall {
-                        setup::waybar::uninstall()?;
-                    } else if json {
-                        println!("{}", setup::waybar::get_json_config());
-                    } else if css {
-                        println!("{}", setup::waybar::get_css_config());
-                    } else {
-                        setup::waybar::print_config();
-                    }
-                }
-                Some(SetupAction::Dms {
-                    install,
-                    uninstall,
-                    qml,
-                }) => {
-                    warn_if_root("dms");
-                    if install {
-                        setup::dms::install()?;
-                    } else if uninstall {
-                        setup::dms::uninstall()?;
-                    } else if qml {
-                        println!("{}", setup::dms::get_qml_config());
-                    } else {
-                        setup::dms::print_config();
-                    }
-                }
-                Some(SetupAction::Model { list, set, restart }) => {
-                    warn_if_root("model");
-                    if list {
-                        setup::model::list_installed();
-                    } else if let Some(model_name) = set {
-                        setup::model::set_model(&model_name, restart).await?;
-                    } else {
-                        setup::model::interactive_select().await?;
-                    }
-                }
-                Some(SetupAction::Gpu {
-                    enable,
-                    disable,
-                    status,
-                }) => {
-                    if status {
-                        setup::gpu::show_status();
-                    } else if enable {
-                        setup::gpu::enable()?;
-                    } else if disable {
-                        setup::gpu::disable()?;
-                    } else {
-                        // Default: show status
-                        setup::gpu::show_status();
-                    }
-                }
-                Some(SetupAction::Variant { to }) => {
-                    let variant =
-                        setup::binary::Variant::from_binary_name(&to).ok_or_else(|| {
-                            anyhow::anyhow!(
-                                "Unknown variant '{}'. Expected one of: {}",
-                                to,
-                                setup::binary::Variant::ALL
-                                    .iter()
-                                    .map(|v| v.binary_name())
-                                    .collect::<Vec<_>>()
-                                    .join(", ")
-                            )
-                        })?;
-                    setup::binary::switch_to(variant)?;
-                    println!("Switched /usr/bin/voxtype to {}.", variant.binary_name());
-                }
-                Some(SetupAction::Onnx {
-                    enable,
-                    disable,
-                    status,
-                })
-                | Some(SetupAction::Parakeet {
-                    enable,
-                    disable,
-                    status,
-                }) => {
-                    warn_if_root("onnx");
-                    if status {
-                        setup::parakeet::show_status();
-                    } else if enable {
-                        setup::parakeet::enable()?;
-                    } else if disable {
-                        setup::parakeet::disable()?;
-                    } else {
-                        // Default: show status
-                        setup::parakeet::show_status();
-                    }
-                }
-                Some(SetupAction::Compositor { compositor_type }) => {
-                    warn_if_root("compositor");
-                    setup::compositor::run(&compositor_type).await?;
-                }
-                Some(SetupAction::Vad { status }) => {
-                    warn_if_root("vad");
-                    if status {
-                        setup::vad::show_status();
-                    } else {
-                        setup::vad::download_model()?;
-                    }
-                }
-                Some(SetupAction::Quickshell {
-                    target,
-                    source,
-                    force,
-                    print_bindings,
-                    bridge,
-                    bridge_target,
-                    skip_bridge,
-                }) => {
-                    warn_if_root("quickshell");
-                    setup::quickshell::run(
-                        target,
-                        source,
-                        force,
-                        print_bindings,
-                        bridge,
-                        bridge_target,
-                        skip_bridge,
-                    )?;
-                }
-                None => {
-                    // Default: run setup (non-blocking)
-                    warn_if_root("");
-                    setup::run_setup(&config, download, model.as_deref(), quiet, no_post_install)
-                        .await?;
-                }
-            }
-        }
-
-        Commands::Config { action } => match action {
-            None => show_config(&config).await?,
-            Some(ConfigAction::Set { key }) => match key {
-                ConfigSetKey::Engine { name } => {
-                    run_config_set_engine(cli.config.clone(), &name)?;
-                }
-            },
-        },
-
-        Commands::Info { action } => {
-            run_info_command(action)?;
-        }
-
-        Commands::Configure { force_package_mode } => {
-            voxtype::tui::run(force_package_mode)?;
-        }
-
-        Commands::Status {
-            follow,
-            format,
-            extended,
-            icon_theme,
-        } => {
-            run_status(&config, follow, &format, extended, icon_theme).await?;
-        }
-
-        Commands::Record { action } => {
-            send_record_command(&config, action, top_level_model.as_deref())?;
-        }
-
-        Commands::Meeting { action } => {
-            run_meeting_command(&config, action).await?;
-        }
-
-        Commands::CheckUpdate => {
-            check_for_updates().await?;
-        }
-    }
-
-    Ok(())
-}
-
-/// First-launch auto-setup for macOS app bundle launches.
-///
-/// Detects if this is the first launch by checking for a config file and downloaded model.
-/// If either is missing, creates default config and downloads the recommended model
-/// so the user can start recording immediately after granting permissions.
-#[cfg(target_os = "macos")]
-async fn first_launch_setup(_config: &config::Config) {
-    // Check if config file exists
-    let config_exists = config::Config::default_path()
-        .map(|p| p.exists())
-        .unwrap_or(false);
-
-    // Check if any model is already downloaded
-    let models_dir = config::Config::models_dir();
-    let has_model = models_dir.exists()
-        && std::fs::read_dir(&models_dir)
-            .map(|entries| {
-                entries.filter_map(|e| e.ok()).any(|e| {
-                    let name = e.file_name().to_string_lossy().to_string();
-                    // Whisper models (ggml-*.bin) or Parakeet/ONNX model dirs
-                    (name.starts_with("ggml-") && name.ends_with(".bin"))
-                        || (e.path().is_dir() && e.path().join("encoder-model.onnx").exists())
-                })
-            })
-            .unwrap_or(false);
-
-    if config_exists && has_model {
-        return; // Not first launch
-    }
-
-    // Create default config if missing
-    if !config_exists {
-        if let Some(config_path) = config::Config::default_path() {
-            if let Some(parent) = config_path.parent() {
-                let _ = std::fs::create_dir_all(parent);
-            }
-            let content = config::default_config_content();
-            if std::fs::write(&config_path, &content).is_ok() {
-                tracing::info!("Created default config: {:?}", config_path);
-            }
-        }
-    }
-
-    // Download default model if none present
-    if !has_model {
-        // Detect system language to choose the right model
-        let is_english = tokio::process::Command::new("defaults")
-            .args(["read", "NSGlobalDomain", "AppleLanguages"])
-            .output()
-            .await
-            .map(|o| {
-                let s = String::from_utf8_lossy(&o.stdout);
-                s.lines()
-                    .find(|l| l.trim().starts_with('"'))
-                    .map(|l| {
-                        l.trim()
-                            .trim_matches(|c| c == '"' || c == ',')
-                            .starts_with("en")
-                    })
-                    .unwrap_or(true)
-            })
-            .unwrap_or(true);
-
-        // Show notification that model is downloading
-        let _ = std::process::Command::new("osascript")
-            .args([
-                "-e",
-                "display notification \"Downloading speech model (this may take a minute)...\" with title \"Voxtype\"",
-            ])
-            .status();
-
-        #[cfg(feature = "parakeet")]
-        let download_result = if is_english {
-            tracing::info!("First launch: downloading Parakeet model");
-            setup::model::download_parakeet_model("parakeet-tdt-0.6b-v3-int8")
-                .and_then(|_| setup::model::set_parakeet_config("parakeet-tdt-0.6b-v3-int8"))
-        } else {
-            tracing::info!("First launch: downloading Whisper base model");
-            setup::model::download_model("base")
-                .and_then(|_| setup::model::set_model_config("base"))
-        };
-
-        #[cfg(not(feature = "parakeet"))]
-        let download_result = {
-            let model = if is_english { "base.en" } else { "base" };
-            tracing::info!("First launch: downloading Whisper {} model", model);
-            setup::model::download_model(model).and_then(|_| setup::model::set_model_config(model))
-        };
-
-        match download_result {
-            Ok(_) => {
-                let _ = std::process::Command::new("osascript")
-                    .args([
-                        "-e",
-                        "display notification \"Ready! Press fn to start recording.\" with title \"Voxtype\"",
-                    ])
-                    .status();
-            }
-            Err(e) => {
-                tracing::error!("Failed to download model: {}", e);
-                let msg = format!(
-                    "display notification \"Model download failed: {}. Run 'voxtype setup model' to try again.\" with title \"Voxtype\"",
-                    e.to_string().replace('"', "'")
-                );
-                let _ = std::process::Command::new("osascript")
-                    .args(["-e", &msg])
-                    .status();
-            }
-        }
-    }
-}
-
-/// Check for updates by comparing version with GitHub releases
-async fn check_for_updates() -> anyhow::Result<()> {
-    let current = env!("CARGO_PKG_VERSION");
-    println!("Voxtype Update Check\n");
-    println!("====================\n");
-    println!("Current version: {}", current);
-    println!("Checking for updates...\n");
-
-    // Fetch latest release from GitHub API (blocking call wrapped in spawn_blocking).
-    // ureq::Error is ~272 bytes; boxing the closure's Result would require an extra
-    // allocation just to satisfy clippy on a one-shot startup-time call.
-    #[allow(clippy::result_large_err)]
-    let result = tokio::task::spawn_blocking(|| {
-        ureq::get("https://api.github.com/repos/peteonrails/voxtype/releases/latest")
-            .set("User-Agent", "voxtype-update-checker")
-            .call()
-    })
-    .await?;
-
-    match result {
-        Ok(resp) => {
-            let release: serde_json::Value = resp.into_json()?;
-            if let Some(tag) = release["tag_name"].as_str() {
-                let latest = tag.trim_start_matches('v');
-
-                // Compare versions using semver
-                let current_ver = semver::Version::parse(current)
-                    .unwrap_or_else(|_| semver::Version::new(0, 0, 0));
-                let latest_ver = semver::Version::parse(latest)
-                    .unwrap_or_else(|_| semver::Version::new(0, 0, 0));
-
-                if latest_ver > current_ver {
-                    println!(
-                        "\x1b[33m⚠ Update available: {} → {}\x1b[0m\n",
-                        current, latest
-                    );
-                    println!(
-                        "Download: https://github.com/peteonrails/voxtype/releases/tag/{}",
-                        tag
-                    );
-                    println!("Website:  https://voxtype.io/download");
-
-                    // Show release notes excerpt if available
-                    if let Some(body) = release["body"].as_str() {
-                        let summary: String = body.lines().take(5).collect::<Vec<_>>().join("\n");
-                        if !summary.is_empty() {
-                            println!("\nRelease notes:");
-                            println!("{}", summary);
-                            if body.lines().count() > 5 {
-                                println!("...");
-                            }
-                        }
-                    }
-                } else {
-                    println!(
-                        "\x1b[32m✓ You're on the latest version ({}).\x1b[0m",
-                        current
-                    );
-                }
-            } else {
-                println!("Could not parse latest version from GitHub.");
-            }
-        }
-        Err(ureq::Error::Status(code, _)) => {
-            eprintln!("GitHub API returned status: {}", code);
-            eprintln!("Try again later or check manually: https://github.com/peteonrails/voxtype/releases");
-        }
-        Err(e) => {
-            eprintln!("Failed to check for updates: {}", e);
-            eprintln!("Check manually: https://github.com/peteonrails/voxtype/releases");
-        }
-    }
-
-    Ok(())
-}
-
-/// Path to the daemon PID file (matches the lockfile the daemon writes via Pidlock).
-///
-/// IMPORTANT: This must match the file used by `send_record_command()` and
-/// `is_daemon_running()`. Historically `check_daemon_running()` read from
-/// `"pid"` while `send_record_command()` read from `"voxtype.lock"`, which
-/// caused `voxtype meeting start/stop/pause/resume` to falsely report the
-/// daemon as not running. See test `pid_file_path_matches_send_record_command`.
-fn daemon_pid_file_path() -> std::path::PathBuf {
-    config::Config::runtime_dir().join("voxtype.lock")
-}
-
-/// Check if the daemon is running, exit with error if not
-fn check_daemon_running() -> anyhow::Result<()> {
-    let pid_file = daemon_pid_file_path();
-
-    if !pid_file.exists() {
-        eprintln!("Error: Voxtype daemon is not running.");
-        eprintln!("Start it with: voxtype daemon");
-        std::process::exit(1);
-    }
-
-    let pid_str = std::fs::read_to_string(&pid_file)
-        .map_err(|e| anyhow::anyhow!("Failed to read PID file: {}", e))?;
-
-    let pid: i32 = pid_str
-        .trim()
-        .parse()
-        .map_err(|e| anyhow::anyhow!("Invalid PID in file: {}", e))?;
-
-    // Check if the process is actually running (signal 0 = check existence)
-    if unsafe { libc::kill(pid, 0) } != 0 {
-        // Process doesn't exist, clean up stale PID file
-        let _ = std::fs::remove_file(&pid_file);
-        eprintln!("Error: Voxtype daemon is not running (stale PID file removed).");
-        eprintln!("Start it with: voxtype daemon");
-        std::process::exit(1);
-    }
-
-    Ok(())
-}
-
-/// Send a record command to the running daemon via Unix signals or file triggers
-fn send_record_command(
-    config: &config::Config,
-    action: RecordAction,
-    top_level_model: Option<&str>,
-) -> anyhow::Result<()> {
-    use voxtype::OutputModeOverride;
-
-    // Read PID from the lock file (daemon writes PID to voxtype.lock)
-    let pid_file = daemon_pid_file_path();
-
-    if !pid_file.exists() {
-        eprintln!("Error: Voxtype daemon is not running.");
-        eprintln!("Start it with: voxtype daemon");
-        std::process::exit(1);
-    }
-
-    let pid_str = std::fs::read_to_string(&pid_file)
-        .map_err(|e| anyhow::anyhow!("Failed to read PID file: {}", e))?;
-
-    let pid: i32 = pid_str
-        .trim()
-        .parse()
-        .map_err(|e| anyhow::anyhow!("Invalid PID in file: {}", e))?;
-
-    // Check if the process is actually running (signal 0 = check existence)
-    if unsafe { libc::kill(pid, 0) } != 0 {
-        // Process doesn't exist, clean up stale PID file
-        let _ = std::fs::remove_file(&pid_file);
-        eprintln!("Error: Voxtype daemon is not running (stale PID file removed).");
-        eprintln!("Start it with: voxtype daemon");
-        std::process::exit(1);
-    }
-
-    // Handle cancel separately (uses file trigger instead of signal)
-    if matches!(action, RecordAction::Cancel) {
-        let cancel_file = config::Config::runtime_dir().join("cancel");
-        std::fs::write(&cancel_file, "cancel")
-            .map_err(|e| anyhow::anyhow!("Failed to write cancel file: {}", e))?;
-        return Ok(());
-    }
-
-    // Write output mode override file if specified
-    // For file mode, format is "file" or "file:/path/to/file"
-    if let Some(mode_override) = action.output_mode_override() {
-        let override_file = config::Config::runtime_dir().join("output_mode_override");
-        let mode_str = match mode_override {
-            OutputModeOverride::Type => "type".to_string(),
-            OutputModeOverride::Clipboard => "clipboard".to_string(),
-            OutputModeOverride::Paste => "paste".to_string(),
-            OutputModeOverride::File => {
-                // Check if explicit path was provided with --file=path
-                match action.file_path() {
-                    Some(path) if !path.is_empty() => format!("file:{}", path),
-                    _ => "file".to_string(),
-                }
-            }
-        };
-        std::fs::write(&override_file, mode_str)
-            .map_err(|e| anyhow::anyhow!("Failed to write output mode override: {}", e))?;
-    }
-
-    // Write model override file if specified (subcommand --model takes priority over top-level --model)
-    let model_override = action.model_override().or(top_level_model);
-    if let Some(model) = model_override {
-        let override_file = config::Config::runtime_dir().join("model_override");
-        std::fs::write(&override_file, model)
-            .map_err(|e| anyhow::anyhow!("Failed to write model override: {}", e))?;
-    }
-
-    // Write smart auto-submit override file if specified
-    if let Some(enabled) = action.smart_auto_submit_override() {
-        let override_file = config::Config::runtime_dir().join("smart_auto_submit_override");
-        std::fs::write(&override_file, if enabled { "true" } else { "false" })
-            .map_err(|e| anyhow::anyhow!("Failed to write smart auto-submit override: {}", e))?;
-    }
-
-    // Write profile override file if specified
-    if let Some(profile_name) = action.profile() {
-        // Validate that the profile exists in config
-        if config.get_profile(profile_name).is_none() {
-            let available = config.profile_names();
-            if available.is_empty() {
-                eprintln!("Error: Profile '{}' not found.", profile_name);
-                eprintln!();
-                eprintln!("No profiles are configured. Add profiles to your config.toml:");
-                eprintln!();
-                eprintln!("  [profiles.{}]", profile_name);
-                eprintln!("  post_process_command = \"your-command-here\"");
-            } else {
-                eprintln!("Error: Profile '{}' not found.", profile_name);
-                eprintln!();
-                eprintln!(
-                    "Available profiles: {}",
-                    available
-                        .iter()
-                        .map(|s| s.as_str())
-                        .collect::<Vec<_>>()
-                        .join(", ")
-                );
-            }
-            std::process::exit(1);
-        }
-
-        let profile_file = config::Config::runtime_dir().join("profile_override");
-        std::fs::write(&profile_file, profile_name)
-            .map_err(|e| anyhow::anyhow!("Failed to write profile override: {}", e))?;
-    }
-
-    // Write auto_submit override file if specified
-    if let Some(value) = action.auto_submit_override() {
-        let override_file = config::Config::runtime_dir().join("auto_submit_override");
-        std::fs::write(&override_file, if value { "true" } else { "false" })
-            .map_err(|e| anyhow::anyhow!("Failed to write auto_submit override: {}", e))?;
-    }
-
-    // Write shift_enter_newlines override file if specified
-    if let Some(value) = action.shift_enter_newlines_override() {
-        let override_file = config::Config::runtime_dir().join("shift_enter_override");
-        std::fs::write(&override_file, if value { "true" } else { "false" })
-            .map_err(|e| anyhow::anyhow!("Failed to write shift_enter override: {}", e))?;
-    }
-
-    // For toggle, we need to read current state to decide which signal to send
-    let signal: libc::c_int = match &action {
-        RecordAction::Start { .. } => libc::SIGUSR1,
-        RecordAction::Stop { .. } => libc::SIGUSR2,
-        RecordAction::Toggle { .. } => {
-            // Read current state to determine action
-            let state_file = match config.resolve_state_file() {
-                Some(path) => path,
-                None => {
-                    eprintln!("Error: Cannot toggle recording without state_file configured.");
-                    eprintln!();
-                    eprintln!("Add to your config.toml:");
-                    eprintln!("  state_file = \"auto\"");
-                    eprintln!();
-                    eprintln!("Or use explicit start/stop commands:");
-                    eprintln!("  voxtype record start");
-                    eprintln!("  voxtype record stop");
-                    std::process::exit(1);
-                }
-            };
-
-            let current_state =
-                std::fs::read_to_string(&state_file).unwrap_or_else(|_| "idle".to_string());
-
-            // "recording" covers the batch and eager paths. "streaming"
-            // covers the Parakeet streaming path. Both are active
-            // capture states whose toggle should send a stop signal,
-            // not start a second session. Without this, toggling
-            // during streaming silently starts a new session while
-            // the original keeps running until the 60s safety
-            // timeout fires — leaking audio into whatever window
-            // has focus.
-            let active = matches!(current_state.trim(), "recording" | "streaming");
-            if active {
-                libc::SIGUSR2 // Stop
-            } else {
-                libc::SIGUSR1 // Start
-            }
-        }
-        RecordAction::Cancel => unreachable!(), // Handled above
-    };
-
-    let result = unsafe { libc::kill(pid, signal) };
-    if result != 0 {
-        return Err(anyhow::anyhow!(
-            "Failed to send signal to daemon: {}",
-            std::io::Error::last_os_error()
-        ));
-    }
-
-    Ok(())
-}
-
-/// Transcribe an audio file
-fn transcribe_file(config: &config::Config, path: &PathBuf) -> anyhow::Result<()> {
-    use hound::WavReader;
-
-    println!("Loading audio file: {:?}", path);
-
-    let reader = WavReader::open(path)?;
-    let spec = reader.spec();
-
-    println!(
-        "Audio format: {} Hz, {} channel(s), {:?}",
-        spec.sample_rate, spec.channels, spec.sample_format
-    );
-
-    // Convert samples to f32 mono at 16kHz
-    let samples: Vec<f32> = match spec.sample_format {
-        hound::SampleFormat::Int => {
-            let max_val = (1 << (spec.bits_per_sample - 1)) as f32;
-            reader
-                .into_samples::<i32>()
-                .filter_map(|s| s.ok())
-                .map(|s| s as f32 / max_val)
-                .collect()
-        }
-        hound::SampleFormat::Float => reader
-            .into_samples::<f32>()
-            .filter_map(|s| s.ok())
-            .collect(),
-    };
-
-    // Mix to mono if stereo
-    let mono_samples: Vec<f32> = if spec.channels > 1 {
-        samples
-            .chunks(spec.channels as usize)
-            .map(|chunk| chunk.iter().sum::<f32>() / chunk.len() as f32)
-            .collect()
-    } else {
-        samples
-    };
-
-    // Resample to 16kHz if needed
-    let final_samples = if spec.sample_rate != 16000 {
-        println!("Resampling from {} Hz to 16000 Hz...", spec.sample_rate);
-        resample(&mono_samples, spec.sample_rate, 16000)
-    } else {
-        mono_samples
-    };
-
-    println!(
-        "Processing {} samples ({:.2}s)...",
-        final_samples.len(),
-        final_samples.len() as f32 / 16000.0
-    );
-
-    // Run VAD if enabled
-    if let Ok(Some(vad)) = vad::create_vad(config) {
-        match vad.detect(&final_samples) {
-            Ok(result) => {
-                println!(
-                    "VAD: {:.2}s speech ({:.1}% of audio)",
-                    result.speech_duration_secs,
-                    result.speech_ratio * 100.0
-                );
-                if !result.has_speech {
-                    println!("No speech detected, skipping transcription.");
-                    return Ok(());
-                }
-            }
-            Err(e) => {
-                eprintln!("VAD warning: {}", e);
-                // Continue with transcription if VAD fails
-            }
-        }
-    }
-
-    // Create transcriber and transcribe
-    let transcriber = transcribe::create_transcriber(config)?;
-    let text = transcriber.transcribe(&final_samples)?;
-
-    println!("\n{}", text);
-    Ok(())
-}
-
-/// Simple linear resampling
-fn resample(samples: &[f32], from_rate: u32, to_rate: u32) -> Vec<f32> {
-    if from_rate == to_rate {
-        return samples.to_vec();
-    }
-
-    let ratio = to_rate as f64 / from_rate as f64;
-    let new_len = (samples.len() as f64 * ratio).ceil() as usize;
-    let mut output = Vec::with_capacity(new_len);
-
-    for i in 0..new_len {
-        let src_idx = i as f64 / ratio;
-        let idx = src_idx.floor() as usize;
-        let frac = (src_idx - idx as f64) as f32;
-
-        let sample = if idx + 1 < samples.len() {
-            samples[idx] * (1.0 - frac) + samples[idx + 1] * frac
-        } else {
-            samples.get(idx).copied().unwrap_or(0.0)
-        };
-
-        output.push(sample);
-    }
-
-    output
-}
-
-/// Extended status info for JSON output
-struct ExtendedStatusInfo {
-    model: String,
-    device: String,
-    backend: String,
-}
-
-impl ExtendedStatusInfo {
-    fn from_config(config: &config::Config) -> Self {
-        // Resolve the actual backend through the inventory machinery — this is
-        // wrapper-script aware (see setup::binary::active_variant) so it
-        // reports correctly whether /usr/bin/voxtype is a plain symlink or the
-        // exec-wrapper used by GPU/ONNX variants. The legacy
-        // setup::gpu::detect_current_backend() path was Whisper-focused: it
-        // would treat any wrapper script as Backend::Native, which is why
-        // waybar previously showed "CPU (native)" while MIGraphX or CUDA was
-        // actually doing the work.
-        let inv = setup::binary::inventory();
-        let backend = if let Some(v) = inv.active_variant {
-            backend_display_for_variant(v).to_string()
-        } else if let Some(b) = setup::gpu::detect_current_backend() {
-            match b {
-                setup::gpu::Backend::Cpu => "CPU (legacy)",
-                setup::gpu::Backend::Native => "CPU (native)",
-                setup::gpu::Backend::Avx2 => "CPU (AVX2)",
-                setup::gpu::Backend::Avx512 => "CPU (AVX-512)",
-                setup::gpu::Backend::Vulkan => "GPU (Vulkan)",
-            }
-            .to_string()
-        } else if let Some(pb) = setup::parakeet::detect_current_parakeet_backend() {
-            pb.display_name().to_string()
-        } else {
-            "unknown".to_string()
-        };
-
-        Self {
-            model: config.model_name().to_string(),
-            device: config.audio.device.clone(),
-            backend,
-        }
-    }
-}
-
-/// User-facing backend label for an active variant. Combines engine family
-/// (Whisper vs ONNX) with the EP/acceleration so both pieces of info land in
-/// waybar tooltips and `voxtype info` output. Whisper variants get a "CPU"/"GPU"
-/// prefix that matches the legacy display strings; ONNX variants spell out the
-/// EP name explicitly so users can tell a CUDA-12 install apart from CUDA-13.
-fn backend_display_for_variant(v: setup::binary::Variant) -> &'static str {
-    use setup::binary::Variant;
-    match v {
-        Variant::WhisperAvx2 => "CPU (AVX2)",
-        Variant::WhisperAvx512 => "CPU (AVX-512)",
-        Variant::WhisperVulkan => "GPU (Vulkan)",
-        Variant::WhisperNative => "CPU (native)",
-        Variant::OnnxAvx2 => "ONNX CPU (AVX2)",
-        Variant::OnnxAvx512 => "ONNX CPU (AVX-512)",
-        Variant::OnnxCuda12 => "ONNX GPU (CUDA 12)",
-        Variant::OnnxCuda13 => "ONNX GPU (CUDA 13)",
-        Variant::OnnxCuda => "ONNX GPU (CUDA)",
-        Variant::OnnxMigraphx => "ONNX GPU (MIGraphX)",
-        Variant::OnnxNative => "ONNX CPU (native)",
-    }
-}
-
-/// Check if the daemon is actually running by verifying the PID file
-fn is_daemon_running() -> bool {
-    // Use the same lockfile path the daemon writes via Pidlock. Reading from
-    // the legacy `pid` file caused `voxtype status` to report `stopped` even
-    // when the daemon was healthy (the daemon now also writes the lockfile
-    // before write_pid_file is reached). See test
-    // `pid_file_path_matches_send_record_command`.
-    let pid_path = daemon_pid_file_path();
-
-    // Read PID from file
-    let pid_str = match std::fs::read_to_string(&pid_path) {
-        Ok(s) => s,
-        Err(_) => return false, // No PID file = not running
-    };
-
-    let pid: i32 = match pid_str.trim().parse() {
-        Ok(p) => p,
-        Err(_) => return false, // Invalid PID = not running
-    };
-
-    // Check if process exists using kill(pid, 0) - works on both Linux and macOS
-    // Signal 0 doesn't send a signal, just checks if process exists and we have permission
-    unsafe { libc::kill(pid, 0) == 0 }
-}
-
-/// Run the status command - show current daemon state
-async fn run_status(
-    config: &config::Config,
-    follow: bool,
-    format: &str,
-    extended: bool,
-    icon_theme_override: Option<String>,
-) -> anyhow::Result<()> {
-    let state_file = config.resolve_state_file();
-
-    if state_file.is_none() {
-        eprintln!("Error: state_file is not configured.");
-        eprintln!();
-        eprintln!("To enable status monitoring, add to your config.toml:");
-        eprintln!();
-        eprintln!("  state_file = \"auto\"");
-        eprintln!();
-        eprintln!("This enables external integrations like Waybar to monitor voxtype state.");
-        std::process::exit(1);
-    }
-
-    let state_path = state_file.unwrap();
-    let ext_info = if extended {
-        Some(ExtendedStatusInfo::from_config(config))
-    } else {
-        None
-    };
-
-    // Use CLI override if provided, otherwise use config
-    let icons = if let Some(ref theme) = icon_theme_override {
-        let mut status_config = config.status.clone();
-        status_config.icon_theme = theme.clone();
-        status_config.resolve_icons()
-    } else {
-        config.status.resolve_icons()
-    };
-
-    if !follow {
-        // One-shot: just read and print current state
-        // First check if daemon is actually running to avoid stale state
-        let state = if !is_daemon_running() {
-            "stopped".to_string()
-        } else {
-            std::fs::read_to_string(&state_path).unwrap_or_else(|_| "stopped".to_string())
-        };
-        let state = state.trim();
-
-        if format == "json" {
-            println!("{}", format_state_json(state, &icons, ext_info.as_ref()));
-        } else {
-            println!("{}", state);
-        }
-        return Ok(());
-    }
-
-    // Follow mode: watch for changes using inotify
-    use notify::{Config as NotifyConfig, RecommendedWatcher, RecursiveMode, Watcher};
-    use std::sync::mpsc::channel;
-    use std::time::Duration;
-
-    // Print initial state (check if daemon is running to avoid stale state)
-    let state = if !is_daemon_running() {
-        "stopped".to_string()
-    } else {
-        std::fs::read_to_string(&state_path).unwrap_or_else(|_| "stopped".to_string())
-    };
-    let state = state.trim();
-    if format == "json" {
-        println!("{}", format_state_json(state, &icons, ext_info.as_ref()));
-    } else {
-        println!("{}", state);
-    }
-
-    // Set up file watcher
-    let (tx, rx) = channel();
-    let mut watcher = RecommendedWatcher::new(
-        move |res| {
-            let _ = tx.send(res);
-        },
-        NotifyConfig::default().with_poll_interval(Duration::from_millis(100)),
-    )?;
-
-    // Watch the state file's parent directory (file may not exist yet)
-    if let Some(parent) = state_path.parent() {
-        std::fs::create_dir_all(parent)?;
-        watcher.watch(parent, RecursiveMode::NonRecursive)?;
-    }
-
-    // Also try to watch the file directly if it exists
-    if state_path.exists() {
-        let _ = watcher.watch(&state_path, RecursiveMode::NonRecursive);
-    }
-
-    let mut last_state = state.to_string();
-
-    loop {
-        match rx.recv_timeout(Duration::from_millis(500)) {
-            Ok(Ok(_event)) => {
-                // File changed, read new state
-                if let Ok(new_state) = std::fs::read_to_string(&state_path) {
-                    let new_state = new_state.trim().to_string();
-                    if new_state != last_state {
-                        if format == "json" {
-                            println!(
-                                "{}",
-                                format_state_json(&new_state, &icons, ext_info.as_ref())
-                            );
-                        } else {
-                            println!("{}", new_state);
-                        }
-                        last_state = new_state;
-                    }
-                }
-            }
-            Ok(Err(e)) => {
-                tracing::warn!("Watch error: {:?}", e);
-            }
-            Err(std::sync::mpsc::RecvTimeoutError::Timeout) => {
-                // Check if daemon stopped (file deleted or process died)
-                if (!state_path.exists() || !is_daemon_running()) && last_state != "stopped" {
-                    if format == "json" {
-                        println!(
-                            "{}",
-                            format_state_json("stopped", &icons, ext_info.as_ref())
-                        );
-                    } else {
-                        println!("stopped");
-                    }
-                    last_state = "stopped".to_string();
-                }
-            }
-            Err(std::sync::mpsc::RecvTimeoutError::Disconnected) => {
-                break;
-            }
-        }
-    }
-
-    Ok(())
-}
-
-/// Format state as JSON for Waybar consumption
-/// The `alt` field enables Waybar's format-icons feature for custom icon mapping
-fn format_state_json(
-    state: &str,
-    icons: &config::ResolvedIcons,
-    extended: Option<&ExtendedStatusInfo>,
-) -> String {
-    let (text, base_tooltip) = match state {
-        "recording" => (&icons.recording, "Recording..."),
-        "streaming" => (&icons.streaming, "Streaming live..."),
-        "transcribing" => (&icons.transcribing, "Transcribing..."),
-        "idle" => (&icons.idle, "Voxtype ready - hold hotkey to record"),
-        "stopped" => (&icons.stopped, "Voxtype not running"),
-        _ => (&icons.idle, "Unknown state"),
-    };
-
-    // alt = state name (for Waybar format-icons mapping)
-    // class = state name (for CSS styling)
-    let alt = state;
-    let class = state;
-
-    match extended {
-        Some(info) => {
-            // Extended format includes model, device, backend
-            let tooltip = format!(
-                "{}\\nModel: {}\\nDevice: {}\\nBackend: {}",
-                base_tooltip, info.model, info.device, info.backend
-            );
-            format!(
-                r#"{{"text": "{}", "alt": "{}", "class": "{}", "tooltip": "{}", "model": "{}", "device": "{}", "backend": "{}"}}"#,
-                text, alt, class, tooltip, info.model, info.device, info.backend
-            )
-        }
-        None => {
-            format!(
-                r#"{{"text": "{}", "alt": "{}", "class": "{}", "tooltip": "{}"}}"#,
-                text, alt, class, base_tooltip
-            )
-        }
-    }
-}
-
-/// Dispatch `voxtype info <subcommand>`.
-fn run_info_command(action: InfoAction) -> anyhow::Result<()> {
-    match action {
-        InfoAction::Variants { json } => {
-            let inv = setup::binary::inventory();
-            if json {
-                println!("{}", serde_json::to_string_pretty(&inv)?);
-            } else {
-                print_variants_text(&inv);
-            }
-        }
-    }
-    Ok(())
-}
-
-fn print_variants_text(inv: &setup::binary::Inventory) {
-    use setup::binary::InstallKind;
-
-    println!("Voxtype install");
-    println!("  Binary:        {}", inv.binary_path.display());
-    println!(
-        "  Install kind:  {}",
-        match inv.install_kind {
-            InstallKind::Package => "package",
-            InstallKind::Source => "source",
-        }
-    );
-    if let Some(dir) = &inv.package_lib_dir {
-        println!("  Lib dir:       {}", dir.display());
-    }
-    if !inv.compiled_features.is_empty() {
-        println!("  Features:      {}", inv.compiled_features.join(", "));
-    }
-
-    println!();
-    println!("Hardware");
-    println!(
-        "  CPU:           AVX2={}, AVX-512={}",
-        inv.cpu.avx2, inv.cpu.avx512
-    );
-    println!(
-        "  GPU:           NVIDIA={}, AMD={}",
-        inv.gpus.nvidia, inv.gpus.amd
-    );
-
-    println!();
-    println!("Recommended for this hardware");
-    println!(
-        "  Whisper:       ★ {}  — {}",
-        inv.recommendation.whisper.display(),
-        inv.recommendation.whisper_reason
-    );
-    println!(
-        "  ONNX:          ★ {}  — {}",
-        inv.recommendation.onnx.display(),
-        inv.recommendation.onnx_reason
-    );
-
-    println!();
-    if matches!(inv.install_kind, InstallKind::Source) {
-        println!("Source build: variant switching not applicable.");
-        println!("To enable a different engine, rebuild with the appropriate Cargo features.");
-        return;
-    }
-
-    println!("Variants");
-    if let Some(active) = inv.active_variant {
-        println!(
-            "  Active:        {} ({})",
-            active.display(),
-            active.binary_name()
-        );
-    } else {
-        println!("  Active:        unknown (symlink missing or unrecognized)");
-    }
-
-    println!();
-    println!("  Available:");
-    for status in &inv.variants {
-        let mark = if status.active {
-            "● active"
-        } else if !status.installed {
-            "  not installed"
-        } else if !status.runs_on_this_cpu {
-            "  installed (won't run on this CPU)"
-        } else if !status.gpu_available {
-            "  installed (no compatible GPU detected)"
-        } else {
-            "  installed"
-        };
-        println!("    {:<22} {}", status.variant.display(), mark);
-    }
-}
-
-/// Show current configuration
-/// Format the `[meeting]` config sections for display in `voxtype config`.
-///
-/// Extracted so the regression test `config_displays_meeting_section`
-/// can assert that the section is present without spawning a process.
-/// The smoke pipeline previously missed that `voxtype config` omitted
-/// the entire `[meeting]` tree even when meeting mode was configured.
-fn format_meeting_config_section(meeting: &config::MeetingConfig) -> String {
-    use std::fmt::Write;
-    let mut s = String::new();
-
-    let _ = writeln!(s, "\n[meeting]");
-    let _ = writeln!(s, "  enabled = {}", meeting.enabled);
-    let _ = writeln!(s, "  chunk_duration_secs = {}", meeting.chunk_duration_secs);
-    let _ = writeln!(s, "  storage_path = {:?}", meeting.storage_path);
-    let _ = writeln!(s, "  retain_audio = {}", meeting.retain_audio);
-    let _ = writeln!(s, "  max_duration_mins = {}", meeting.max_duration_mins);
-
-    let _ = writeln!(s, "\n[meeting.audio]");
-    let _ = writeln!(s, "  mic_device = {:?}", meeting.audio.mic_device);
-    let _ = writeln!(s, "  loopback_device = {:?}", meeting.audio.loopback_device);
-    let _ = writeln!(s, "  echo_cancel = {:?}", meeting.audio.echo_cancel);
-
-    let _ = writeln!(s, "\n[meeting.diarization]");
-    let _ = writeln!(s, "  enabled = {}", meeting.diarization.enabled);
-    let _ = writeln!(s, "  backend = {:?}", meeting.diarization.backend);
-    let _ = writeln!(s, "  max_speakers = {}", meeting.diarization.max_speakers);
-    if let Some(ref path) = meeting.diarization.model_path {
-        let _ = writeln!(s, "  model_path = {:?}", path);
-    }
-    let _ = writeln!(
-        s,
-        "  min_segment_ms = {}",
-        meeting.diarization.min_segment_ms
-    );
-
-    let _ = writeln!(s, "\n[meeting.summary]");
-    let _ = writeln!(s, "  backend = {:?}", meeting.summary.backend);
-    let _ = writeln!(s, "  ollama_url = {:?}", meeting.summary.ollama_url);
-    let _ = writeln!(s, "  ollama_model = {:?}", meeting.summary.ollama_model);
-    if let Some(ref endpoint) = meeting.summary.remote_endpoint {
-        let _ = writeln!(s, "  remote_endpoint = {:?}", endpoint);
-    }
-    if meeting.summary.remote_api_key.is_some() {
-        let _ = writeln!(s, "  remote_api_key = (set)");
-    }
-    let _ = writeln!(s, "  timeout_secs = {}", meeting.summary.timeout_secs);
-
-    s
-}
-
-/// Resolve the config file path the same way the daemon does — honoring
-/// `--config <FILE>` first, then the existing user/system path, then the
-/// XDG default. The default path is used even when nothing is on disk so
-/// the file gets created in a predictable location on first write.
-fn resolve_config_path_for_write(cli_override: Option<PathBuf>) -> anyhow::Result<PathBuf> {
-    if let Some(p) = cli_override {
-        return Ok(p);
-    }
-    if let Some(p) = config::Config::resolve_existing_path() {
-        return Ok(p);
-    }
-    config::Config::default_path().ok_or_else(|| {
-        anyhow::anyhow!(
-            "Cannot determine config path. Set $XDG_CONFIG_HOME or $HOME, \
-             or pass --config <FILE>."
-        )
-    })
-}
-
-/// Dispatcher for `voxtype config set engine <NAME>`. Exits the process
-/// with code 2 on validation errors (bad name or missing feature) and code
-/// 1 on filesystem failures, matching the contract documented in
-/// `voxtype config set --help`.
-fn run_config_set_engine(cli_override: Option<PathBuf>, name: &str) -> anyhow::Result<()> {
-    let path = resolve_config_path_for_write(cli_override)?;
-    match config_set::set_engine(path, name) {
-        Ok(written) => {
-            println!("Set engine = \"{}\" in {}", name, written.display());
-            println!("Restart voxtype to apply: systemctl --user restart voxtype");
-            Ok(())
-        }
-        Err(e @ config_set::ConfigSetError::UnknownEngine(_))
-        | Err(e @ config_set::ConfigSetError::FeatureNotCompiled(_)) => {
-            eprintln!("error: {}", e);
-            std::process::exit(2);
-        }
-        Err(e @ config_set::ConfigSetError::Editor(_)) => {
-            eprintln!("error: {}", e);
-            std::process::exit(1);
-        }
-    }
-}
-
-async fn show_config(config: &config::Config) -> anyhow::Result<()> {
-    println!("Current Configuration\n");
-    println!("=====================\n");
-
-    println!("[hotkey]");
-    println!("  key = {:?}", config.hotkey.key);
-    println!("  modifiers = {:?}", config.hotkey.modifiers);
-    println!("  mode = {:?}", config.hotkey.mode);
-
-    println!("\n[audio]");
-    println!("  device = {:?}", config.audio.device);
-    println!("  sample_rate = {}", config.audio.sample_rate);
-    println!("  max_duration_secs = {}", config.audio.max_duration_secs);
-
-    println!("\n[audio.feedback]");
-    println!("  enabled = {}", config.audio.feedback.enabled);
-    println!("  theme = {:?}", config.audio.feedback.theme);
-    println!("  volume = {}", config.audio.feedback.volume);
-
-    // Show current engine
-    println!("\n[engine]");
-    println!("  engine = {:?}", config.engine);
-
-    println!("\n[whisper]");
-    println!("  model = {:?}", config.whisper.model);
-    println!("  language = {:?}", config.whisper.language);
-    println!("  translate = {}", config.whisper.translate);
-    if let Some(threads) = config.whisper.threads {
-        println!("  threads = {}", threads);
-    }
-    if let Some(gpu_device) = config.whisper.gpu_device {
-        println!("  gpu_device = {}", gpu_device);
-    }
-
-    // Show Parakeet status
-    println!("\n[parakeet]");
-    if let Some(ref parakeet_config) = config.parakeet {
-        println!("  model = {:?}", parakeet_config.model);
-        if let Some(ref model_type) = parakeet_config.model_type {
-            println!("  model_type = {:?}", model_type);
-        }
-        println!(
-            "  on_demand_loading = {}",
-            parakeet_config.on_demand_loading
-        );
-    } else {
-        println!("  (not configured)");
-    }
-
-    // Check for available Parakeet models
-    let models_dir = config::Config::models_dir();
-    let mut parakeet_models: Vec<String> = Vec::new();
-    if let Ok(entries) = std::fs::read_dir(&models_dir) {
-        for entry in entries.flatten() {
-            let path = entry.path();
-            if path.is_dir() {
-                let name = entry.file_name().to_string_lossy().to_string();
-                if name.contains("parakeet") {
-                    // Check if it has the required ONNX files
-                    let has_encoder = path.join("encoder-model.onnx").exists();
-                    let has_decoder = path.join("decoder_joint-model.onnx").exists()
-                        || path.join("model.onnx").exists();
-                    if has_encoder || has_decoder {
-                        parakeet_models.push(name);
-                    }
-                }
-            }
-        }
-    }
-    if parakeet_models.is_empty() {
-        println!("  available models: (none found)");
-    } else {
-        println!("  available models: {}", parakeet_models.join(", "));
-    }
-
-    // Show Moonshine status (experimental)
-    println!("\n[moonshine] (EXPERIMENTAL)");
-    if let Some(ref moonshine_config) = config.moonshine {
-        println!("  model = {:?}", moonshine_config.model);
-        println!("  quantized = {}", moonshine_config.quantized);
-        if let Some(threads) = moonshine_config.threads {
-            println!("  threads = {}", threads);
-        }
-        println!(
-            "  on_demand_loading = {}",
-            moonshine_config.on_demand_loading
-        );
-    } else {
-        println!("  (not configured)");
-    }
-
-    // Check for available Moonshine models
-    let mut moonshine_models: Vec<String> = Vec::new();
-    if let Ok(entries) = std::fs::read_dir(&models_dir) {
-        for entry in entries.flatten() {
-            let path = entry.path();
-            if path.is_dir() {
-                let name = entry.file_name().to_string_lossy().to_string();
-                if name.contains("moonshine") {
-                    let has_encoder = path.join("encoder_model.onnx").exists()
-                        || path.join("encoder_model_quantized.onnx").exists();
-                    let has_decoder = path.join("decoder_model_merged.onnx").exists()
-                        || path.join("decoder_model_merged_quantized.onnx").exists();
-                    if has_encoder || has_decoder {
-                        moonshine_models.push(name);
-                    }
-                }
-            }
-        }
-    }
-    if moonshine_models.is_empty() {
-        println!("  available models: (none found)");
-    } else {
-        println!("  available models: {}", moonshine_models.join(", "));
-    }
-
-    // Show SenseVoice status (experimental)
-    println!("\n[sensevoice] (EXPERIMENTAL)");
-    if let Some(ref sensevoice_config) = config.sensevoice {
-        println!("  model = {:?}", sensevoice_config.model);
-        println!("  language = {:?}", sensevoice_config.language);
-        println!("  use_itn = {}", sensevoice_config.use_itn);
-        if let Some(threads) = sensevoice_config.threads {
-            println!("  threads = {}", threads);
-        }
-        println!(
-            "  on_demand_loading = {}",
-            sensevoice_config.on_demand_loading
-        );
-    } else {
-        println!("  (not configured)");
-    }
-
-    // Check for available SenseVoice models
-    let mut sensevoice_models: Vec<String> = Vec::new();
-    if let Ok(entries) = std::fs::read_dir(&models_dir) {
-        for entry in entries.flatten() {
-            let path = entry.path();
-            if path.is_dir() {
-                let name = entry.file_name().to_string_lossy().to_string();
-                if name.contains("sensevoice") {
-                    let has_model =
-                        path.join("model.int8.onnx").exists() || path.join("model.onnx").exists();
-                    let has_tokens = path.join("tokens.txt").exists();
-                    if has_model && has_tokens {
-                        sensevoice_models.push(name);
-                    }
-                }
-            }
-        }
-    }
-    if sensevoice_models.is_empty() {
-        println!("  available models: (none found)");
-    } else {
-        println!("  available models: {}", sensevoice_models.join(", "));
-    }
-
-    println!("\n[output]");
-    println!("  mode = {:?}", config.output.mode);
-    println!(
-        "  fallback_to_clipboard = {}",
-        config.output.fallback_to_clipboard
-    );
-    if let Some(ref driver_order) = config.output.driver_order {
-        println!(
-            "  driver_order = [{}]",
-            driver_order
-                .iter()
-                .map(|d| format!("{:?}", d))
-                .collect::<Vec<_>>()
-                .join(", ")
-        );
-    } else {
-        println!("  driver_order = (default: wtype -> dotool -> ydotool -> clipboard)");
-    }
-    println!("  type_delay_ms = {}", config.output.type_delay_ms);
-    println!("  pre_type_delay_ms = {}", config.output.pre_type_delay_ms);
-    println!("  restore_clipboard = {}", config.output.restore_clipboard);
-    println!(
-        "  restore_clipboard_delay_ms = {}",
-        config.output.restore_clipboard_delay_ms
-    );
-    println!(
-        "  wait_for_modifier_release = {}",
-        config.output.wait_for_modifier_release
-    );
-    println!(
-        "  modifier_release_timeout_ms = {}",
-        config.output.modifier_release_timeout_ms
-    );
-
-    println!("\n[output.notification]");
-    println!(
-        "  on_recording_start = {}",
-        config.output.notification.on_recording_start
-    );
-    println!(
-        "  on_recording_stop = {}",
-        config.output.notification.on_recording_stop
-    );
-    println!(
-        "  on_transcription = {}",
-        config.output.notification.on_transcription
-    );
-    println!("  urgency = {:?}", config.output.notification.urgency);
-
-    // Meeting mode is opt-in. Always show the section so users can see the
-    // resolved defaults — the smoke runner previously missed this entirely
-    // and meeting users had no way to verify their config was loaded.
-    print!("{}", format_meeting_config_section(&config.meeting));
-
-    println!("\n[status]");
-    println!("  icon_theme = {:?}", config.status.icon_theme);
-    let icons = config.status.resolve_icons();
-    println!(
-        "  (resolved icons: idle={:?} recording={:?} transcribing={:?} stopped={:?})",
-        icons.idle, icons.recording, icons.transcribing, icons.stopped
-    );
-
-    if let Some(ref state_file) = config.state_file {
-        println!("\n[integration]");
-        println!("  state_file = {:?}", state_file);
-        if let Some(resolved) = config.resolve_state_file() {
-            println!("  (resolves to: {:?})", resolved);
-        }
-    }
-
-    // Show output chain status
-    let output_status = setup::detect_output_chain().await;
-    setup::print_output_chain_status(&output_status);
-
-    println!("\n---");
-    match config::Config::resolve_existing_path() {
-        Some(path) => println!("Config file: {:?} (loaded)", path),
-        None => println!(
-            "Config file: {:?} (not found, using defaults; system fallback {:?} also missing)",
-            config::Config::default_path().unwrap_or_else(|| PathBuf::from("(unknown)")),
-            config::Config::system_path()
-        ),
-    }
-    println!("Models dir: {:?}", config::Config::models_dir());
-
-    Ok(())
-}
-
-/// Reset SIGPIPE to default behavior (terminate process) instead of the Rust
-/// default of ignoring it. This prevents panics when stdout is piped through
-/// commands like `head` that close the pipe early.
-#[cfg(unix)]
-fn reset_sigpipe() {
-    unsafe {
-        libc::signal(libc::SIGPIPE, libc::SIG_DFL);
-    }
-}
-
-#[cfg(not(unix))]
-fn reset_sigpipe() {
-    // No-op on non-Unix platforms
-}
-
-/// Run a meeting command
-async fn run_meeting_command(config: &config::Config, action: MeetingAction) -> anyhow::Result<()> {
-    use meeting::{export_meeting, ExportFormat, ExportOptions, MeetingConfig, StorageConfig};
-
-    // Convert config to meeting config
-    let storage_path = if config.meeting.storage_path == "auto" {
-        StorageConfig::default_storage_path()
-    } else {
-        PathBuf::from(&config.meeting.storage_path)
-    };
-
-    let meeting_config = MeetingConfig {
-        enabled: config.meeting.enabled,
-        chunk_duration_secs: config.meeting.chunk_duration_secs,
-        storage: StorageConfig {
-            storage_path,
-            retain_audio: config.meeting.retain_audio,
-            max_meetings: 0,
-        },
-        retain_audio: config.meeting.retain_audio,
-        max_duration_mins: config.meeting.max_duration_mins,
-        vad_threshold: config.meeting.audio.vad_threshold,
-        diarization: None,
-    };
-
-    match action {
-        MeetingAction::Start { title, diarization } => {
-            // Check if meeting mode is enabled
-            if !config.meeting.enabled {
-                eprintln!("Error: Meeting mode is disabled in config.");
-                eprintln!();
-                eprintln!("Enable it by adding to config.toml:");
-                eprintln!("  [meeting]");
-                eprintln!("  enabled = true");
-                std::process::exit(1);
-            }
-
-            // Check if daemon is running
-            check_daemon_running()?;
-
-            // Check if meeting already in progress
-            let meeting_state_file = config::Config::runtime_dir().join("meeting_state");
-            if meeting_state_file.exists() {
-                let state = std::fs::read_to_string(&meeting_state_file).unwrap_or_default();
-                if state.starts_with("recording") || state.starts_with("paused") {
-                    eprintln!("Error: A meeting is already in progress.");
-                    eprintln!("Use 'voxtype meeting stop' to end it first.");
-                    std::process::exit(1);
-                }
-            }
-
-            // --diarization ml requires the ml-diarization feature at build
-            // time. Without it the daemon's diarizer factory silently falls
-            // back, leaving the CLI's "(diarization backend: ml)" exit
-            // message a lie. Reject the request up front with a pointer to
-            // the binaries that DO carry ml-diarization.
-            if diarization.as_deref() == Some("ml") && !cfg!(feature = "ml-diarization") {
-                eprintln!("Error: --diarization ml requested but this binary was not built with");
-                eprintln!(
-                    "  the `ml-diarization` feature. ECAPA-TDNN diarization is shipped in the"
-                );
-                eprintln!(
-                    "  ONNX binaries (voxtype-onnx-avx2, voxtype-onnx-avx512, voxtype-onnx-cuda-*,"
-                );
-                eprintln!(
-                    "  voxtype-onnx-migraphx). Install one of those, or omit --diarization to"
-                );
-                eprintln!("  use the source-based `simple` backend.");
-                std::process::exit(1);
-            }
-
-            // Ensure GTCRN speech enhancement model is available
-            setup::model::ensure_gtcrn_model();
-
-            // A --diarization override only changes the *backend*; it cannot
-            // turn diarization on when config has disabled it. Warn loudly so
-            // users don't think they're getting speaker labels they aren't.
-            let diarization_active = config.meeting.diarization.enabled;
-            if diarization.is_some() && !diarization_active {
-                eprintln!(
-                    "Warning: --diarization is a backend override and only takes effect when"
-                );
-                eprintln!(
-                    "  [meeting.diarization] enabled = true in config; diarization is disabled,"
-                );
-                eprintln!("  so the override will be ignored for this meeting.");
-            }
-
-            // Write the diarization override first so it's visible by the time
-            // the daemon picks up the start trigger.
-            let runtime_dir = config::Config::runtime_dir();
-            let diarization_file = runtime_dir.join("meeting_start_diarization");
-            if let Some(ref backend) = diarization {
-                std::fs::write(&diarization_file, backend)?;
-            } else {
-                // Clear any stale override left from a prior run.
-                let _ = std::fs::remove_file(&diarization_file);
-            }
-
-            // Write start trigger file (with optional title)
-            let start_file = runtime_dir.join("meeting_start");
-            let content = title.unwrap_or_default();
-            std::fs::write(&start_file, content)?;
-
-            let suffix = diarization
-                .as_deref()
-                .filter(|_| diarization_active)
-                .map(|b| format!(" (diarization backend: {})", b))
-                .unwrap_or_default();
-            println!(
-                "Meeting start requested{}. Check status with 'voxtype meeting status'.",
-                suffix
-            );
-        }
-
-        MeetingAction::Stop => {
-            check_daemon_running()?;
-
-            // Check if meeting is in progress
-            let meeting_state_file = config::Config::runtime_dir().join("meeting_state");
-            if !meeting_state_file.exists() {
-                eprintln!("Error: No meeting in progress.");
-                std::process::exit(1);
-            }
-
-            let state = std::fs::read_to_string(&meeting_state_file).unwrap_or_default();
-            if state.starts_with("idle") || state.is_empty() {
-                eprintln!("Error: No meeting in progress.");
-                std::process::exit(1);
-            }
-
-            // Write stop trigger file
-            let stop_file = config::Config::runtime_dir().join("meeting_stop");
-            std::fs::write(&stop_file, "")?;
-
-            println!("Meeting stop requested.");
-        }
-
-        MeetingAction::Pause => {
-            check_daemon_running()?;
-
-            // Check if meeting is active (not paused)
-            let meeting_state_file = config::Config::runtime_dir().join("meeting_state");
-            if !meeting_state_file.exists() {
-                eprintln!("Error: No meeting in progress.");
-                std::process::exit(1);
-            }
-
-            let state = std::fs::read_to_string(&meeting_state_file).unwrap_or_default();
-            if !state.starts_with("recording") {
-                eprintln!("Error: No active meeting to pause.");
-                std::process::exit(1);
-            }
-
-            // Write pause trigger file
-            let pause_file = config::Config::runtime_dir().join("meeting_pause");
-            std::fs::write(&pause_file, "")?;
-
-            println!("Meeting pause requested.");
-        }
-
-        MeetingAction::Resume => {
-            check_daemon_running()?;
-
-            // Check if meeting is paused
-            let meeting_state_file = config::Config::runtime_dir().join("meeting_state");
-            if !meeting_state_file.exists() {
-                eprintln!("Error: No paused meeting to resume.");
-                std::process::exit(1);
-            }
-
-            let state = std::fs::read_to_string(&meeting_state_file).unwrap_or_default();
-            if !state.starts_with("paused") {
-                eprintln!("Error: No paused meeting to resume.");
-                std::process::exit(1);
-            }
-
-            // Write resume trigger file
-            let resume_file = config::Config::runtime_dir().join("meeting_resume");
-            std::fs::write(&resume_file, "")?;
-
-            println!("Meeting resume requested.");
-        }
-
-        MeetingAction::Status => {
-            // Read meeting state file
-            let meeting_state_file = config::Config::runtime_dir().join("meeting_state");
-            if !meeting_state_file.exists() {
-                println!("No meeting currently in progress.");
-                println!();
-                println!("Use 'voxtype meeting list' to see past meetings.");
-                return Ok(());
-            }
-
-            let state = std::fs::read_to_string(&meeting_state_file).unwrap_or_default();
-            let lines: Vec<&str> = state.lines().collect();
-
-            if lines.is_empty() || lines[0] == "idle" {
-                println!("No meeting currently in progress.");
-                println!();
-                println!("Use 'voxtype meeting list' to see past meetings.");
-            } else {
-                let status = lines[0];
-                let meeting_id = lines.get(1).unwrap_or(&"");
-
-                println!("Meeting Status: {}", status);
-                if !meeting_id.is_empty() {
-                    println!("Meeting ID: {}", meeting_id);
-                }
-            }
-        }
-
-        MeetingAction::List { limit } => {
-            match meeting::list_meetings(&meeting_config, Some(limit)) {
-                Ok(meetings) => {
-                    if meetings.is_empty() {
-                        println!("No meetings found.");
-                        return Ok(());
-                    }
-
-                    println!("Recent Meetings");
-                    println!("===============\n");
-
-                    for m in meetings {
-                        let duration = m
-                            .duration_secs
-                            .map(|d| {
-                                let mins = d / 60;
-                                let secs = d % 60;
-                                format!("{}m {}s", mins, secs)
-                            })
-                            .unwrap_or_else(|| "in progress".to_string());
-
-                        println!("{}", m.display_title());
-                        println!("  ID: {}", m.id);
-                        println!("  Date: {}", m.started_at.format("%Y-%m-%d %H:%M"));
-                        println!("  Duration: {}", duration);
-                        println!("  Status: {:?}", m.status);
-                        println!();
-                    }
-                }
-                Err(e) => {
-                    eprintln!("Error listing meetings: {}", e);
-                    std::process::exit(1);
-                }
-            }
-        }
-
-        MeetingAction::Export {
-            meeting_id,
-            format,
-            output,
-            timestamps,
-            speakers,
-            metadata,
-        } => {
-            let export_format = ExportFormat::parse(&format).ok_or_else(|| {
-                anyhow::anyhow!(
-                    "Unknown export format '{}'. Valid formats: text, markdown, json",
-                    format
-                )
-            })?;
-
-            let options = ExportOptions {
-                include_timestamps: timestamps,
-                include_speakers: speakers,
-                include_metadata: metadata,
-                line_width: 0,
-            };
-
-            let meeting_data = match meeting::get_meeting(&meeting_config, &meeting_id) {
-                Ok(m) => m,
-                Err(e) => {
-                    eprintln!("Error loading meeting: {}", e);
-                    std::process::exit(1);
-                }
-            };
-
-            let content = match export_meeting(&meeting_data, export_format, &options) {
-                Ok(c) => c,
-                Err(e) => {
-                    eprintln!("Error exporting meeting: {}", e);
-                    std::process::exit(1);
-                }
-            };
-
-            if let Some(path) = output {
-                let file_path = if path.is_dir() {
-                    let title = meeting_data.metadata.display_title();
-                    let safe_title =
-                        title.replace(['/', '\\', ':', '*', '?', '"', '<', '>', '|'], "-");
-                    let basename = if safe_title.trim().is_empty() {
-                        format!("meeting-{}", meeting_data.metadata.id)
-                    } else {
-                        safe_title
-                    };
-                    path.join(format!("{}.{}", basename, export_format.extension()))
-                } else {
-                    path
-                };
-                std::fs::write(&file_path, &content)?;
-                println!("Exported to {}", file_path.display());
-            } else {
-                println!("{}", content);
-            }
-        }
-
-        MeetingAction::Show { meeting_id } => {
-            match meeting::get_meeting(&meeting_config, &meeting_id) {
-                Ok(meeting) => {
-                    println!("{}", meeting.metadata.display_title());
-                    println!("{}", "=".repeat(meeting.metadata.display_title().len()));
-                    println!();
-                    println!("ID:       {}", meeting.metadata.id);
-                    println!(
-                        "Started:  {}",
-                        meeting.metadata.started_at.format("%Y-%m-%d %H:%M UTC")
-                    );
-                    if let Some(ended) = meeting.metadata.ended_at {
-                        println!("Ended:    {}", ended.format("%Y-%m-%d %H:%M UTC"));
-                    }
-                    if let Some(duration) = meeting.metadata.duration_secs {
-                        let hours = duration / 3600;
-                        let mins = (duration % 3600) / 60;
-                        let secs = duration % 60;
-                        if hours > 0 {
-                            println!("Duration: {}h {}m {}s", hours, mins, secs);
-                        } else {
-                            println!("Duration: {}m {}s", mins, secs);
-                        }
-                    }
-                    println!("Status:   {:?}", meeting.metadata.status);
-                    println!("Chunks:   {}", meeting.metadata.chunk_count);
-                    println!();
-                    println!("Transcript:");
-                    println!("-----------");
-                    println!("Segments: {}", meeting.transcript.segments.len());
-                    println!("Words:    {}", meeting.transcript.word_count());
-                    println!("Speakers: {}", meeting.transcript.speakers().join(", "));
-                    println!();
-                    println!(
-                        "Use 'voxtype meeting export {}' to export the transcript.",
-                        meeting_id
-                    );
-                }
-                Err(e) => {
-                    eprintln!("Error loading meeting: {}", e);
-                    std::process::exit(1);
-                }
-            }
-        }
-
-        MeetingAction::Delete { meeting_id, force } => {
-            if !force {
-                eprintln!("This will permanently delete the meeting and all associated files.");
-                eprintln!("Use --force to confirm deletion.");
-                std::process::exit(1);
-            }
-
-            let storage = meeting::MeetingStorage::open(meeting_config.storage.clone())
-                .map_err(|e| anyhow::anyhow!("Failed to open storage: {}", e))?;
-
-            let id = storage
-                .resolve_meeting_id(&meeting_id)
-                .map_err(|e| anyhow::anyhow!("Meeting not found: {}", e))?;
-
-            storage
-                .delete_meeting(&id)
-                .map_err(|e| anyhow::anyhow!("Failed to delete meeting: {}", e))?;
-
-            println!("Meeting {} deleted.", meeting_id);
-        }
-
-        MeetingAction::Label {
-            meeting_id,
-            speaker_id,
-            label,
-        } => {
-            let storage = meeting::MeetingStorage::open(meeting_config.storage.clone())
-                .map_err(|e| anyhow::anyhow!("Failed to open storage: {}", e))?;
-
-            let id = storage
-                .resolve_meeting_id(&meeting_id)
-                .map_err(|e| anyhow::anyhow!("Meeting not found: {}", e))?;
-
-            // Parse speaker_id - accept "SPEAKER_00", "0", "00", etc.
-            let speaker_num: u32 = if speaker_id.starts_with("SPEAKER_") {
-                speaker_id
-                    .trim_start_matches("SPEAKER_")
-                    .parse()
-                    .map_err(|_| anyhow::anyhow!("Invalid speaker ID format: {}", speaker_id))?
-            } else {
-                speaker_id.parse().map_err(|_| {
-                    anyhow::anyhow!(
-                        "Invalid speaker ID: {}. Use SPEAKER_XX or a number.",
-                        speaker_id
-                    )
-                })?
-            };
-
-            storage
-                .set_speaker_label(&id, speaker_num, &label)
-                .map_err(|e| anyhow::anyhow!("Failed to set speaker label: {}", e))?;
-
-            println!(
-                "Labeled SPEAKER_{:02} as '{}' in meeting {}",
-                speaker_num, label, meeting_id
-            );
-        }
-
-        MeetingAction::Summarize {
-            meeting_id,
-            format,
-            output,
-        } => {
-            // Load meeting
-            let meeting = meeting::get_meeting(&meeting_config, &meeting_id)
-                .map_err(|e| anyhow::anyhow!("Failed to load meeting: {}", e))?;
-
-            // Create summary config from meeting config
-            let summary_config = meeting::summary::SummaryConfig {
-                backend: config.meeting.summary.backend.clone(),
-                ollama_url: config.meeting.summary.ollama_url.clone(),
-                ollama_model: config.meeting.summary.ollama_model.clone(),
-                remote_endpoint: config.meeting.summary.remote_endpoint.clone(),
-                remote_api_key: config.meeting.summary.remote_api_key.clone(),
-                timeout_secs: config.meeting.summary.timeout_secs,
-            };
-
-            // Create summarizer
-            let summarizer = meeting::summary::create_summarizer(&summary_config)
-                .ok_or_else(|| {
-                    anyhow::anyhow!(
-                        "Summarization not configured. Set [meeting.summary] backend in config.toml:\n\n\
-                        [meeting.summary]\n\
-                        backend = \"local\"  # or \"remote\"\n\
-                        ollama_url = \"http://localhost:11434\"\n\
-                        ollama_model = \"llama3.2\""
-                    )
-                })?;
-
-            // Check availability
-            if !summarizer.is_available() {
-                return Err(anyhow::anyhow!(
-                    "Summarizer '{}' is not available. Check that Ollama is running.",
-                    summarizer.name()
-                ));
-            }
-
-            eprintln!("Generating summary using {}...", summarizer.name());
-
-            // Generate summary
-            let summary = summarizer
-                .summarize(&meeting)
-                .map_err(|e| anyhow::anyhow!("Summarization failed: {}", e))?;
-
-            // Format output
-            let content = match format.as_str() {
-                "json" => serde_json::to_string_pretty(&summary)
-                    .map_err(|e| anyhow::anyhow!("Failed to serialize summary: {}", e))?,
-                "text" => {
-                    let mut text = String::new();
-                    text.push_str(&format!("Summary: {}\n\n", summary.summary));
-
-                    if !summary.key_points.is_empty() {
-                        text.push_str("Key Points:\n");
-                        for point in &summary.key_points {
-                            text.push_str(&format!("  - {}\n", point));
-                        }
-                        text.push('\n');
-                    }
-
-                    if !summary.action_items.is_empty() {
-                        text.push_str("Action Items:\n");
-                        for item in &summary.action_items {
-                            let assignee = item
-                                .assignee
-                                .as_ref()
-                                .map(|a| format!(" ({})", a))
-                                .unwrap_or_default();
-                            text.push_str(&format!("  - {}{}\n", item.description, assignee));
-                        }
-                        text.push('\n');
-                    }
-
-                    if !summary.decisions.is_empty() {
-                        text.push_str("Decisions:\n");
-                        for decision in &summary.decisions {
-                            text.push_str(&format!("  - {}\n", decision));
-                        }
-                    }
-
-                    text
-                }
-                _ => meeting::summary::summary_to_markdown(&summary),
-            };
-
-            // Output
-            if let Some(path) = output {
-                std::fs::write(&path, &content)?;
-                eprintln!("Summary saved to {:?}", path);
-            } else {
-                println!("{}", content);
-            }
-        }
-    }
-
-    Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    /// Regression: `check_daemon_running()` once read `runtime_dir/pid` while
-    /// `send_record_command()` read `runtime_dir/voxtype.lock`. The mismatch
-    /// caused `voxtype meeting start/stop/pause/resume` to falsely report
-    /// "daemon not running" even when the daemon was healthy. Both helpers
-    /// (and `is_daemon_running`) must resolve to the same path.
-    #[test]
-    fn pid_file_path_matches_send_record_command() {
-        let canonical = daemon_pid_file_path();
-
-        // Sanity: the canonical path ends in `voxtype.lock` (the Pidlock file
-        // the daemon actually writes), not the legacy `pid` filename.
-        assert!(
-            canonical.ends_with("voxtype.lock"),
-            "daemon_pid_file_path() must point at voxtype.lock so meeting \
-             and record commands agree with the daemon's Pidlock. Got: {:?}",
-            canonical,
-        );
-
-        // Both helpers must resolve to exactly the same path. If you split
-        // them, `voxtype meeting start` will regress in the same way it did
-        // before this test existed.
-        let from_send = config::Config::runtime_dir().join("voxtype.lock");
-        assert_eq!(canonical, from_send);
-    }
-
-    /// Regression: after `voxtype record cancel` the daemon writes "idle"
-    /// to the state file. `format_state_json` must render "idle" as the
-    /// idle icon, NOT downgrade or alias it to "stopped". "stopped" is
-    /// reserved for "daemon process not running" (state file missing).
-    #[test]
-    fn record_cancel_leaves_idle_not_stopped() {
-        let icons = config::StatusConfig::default().resolve_icons();
-        let json = format_state_json("idle", &icons, None);
-        assert!(
-            json.contains("\"alt\": \"idle\""),
-            "format_state_json('idle') must keep alt=idle so Waybar shows \
-             the idle icon after `record cancel`. Got: {}",
-            json
-        );
-        assert!(
-            !json.contains("\"alt\": \"stopped\""),
-            "format_state_json('idle') must not alias to 'stopped'. Got: {}",
-            json
-        );
-
-        // And stopped should still map distinctly so we don't accidentally
-        // collapse the two states in the other direction.
-        let stopped_json = format_state_json("stopped", &icons, None);
-        assert!(stopped_json.contains("\"alt\": \"stopped\""));
-    }
-
-    /// Regression: `voxtype config` once omitted the entire `[meeting]`
-    /// tree, leaving meeting-mode users with no way to verify their config
-    /// loaded correctly. The helper must emit all four sub-sections so the
-    /// smoke pipeline can grep for them.
-    #[test]
-    fn config_displays_meeting_section() {
-        let meeting = config::MeetingConfig::default();
-        let rendered = format_meeting_config_section(&meeting);
-
-        for header in [
-            "[meeting]",
-            "[meeting.audio]",
-            "[meeting.diarization]",
-            "[meeting.summary]",
-        ] {
-            assert!(
-                rendered.contains(header),
-                "format_meeting_config_section must emit {} header. Got:\n{}",
-                header,
-                rendered
-            );
-        }
-
-        // A couple of representative fields, so renaming a struct member
-        // doesn't silently drop output without breaking the test.
-        assert!(rendered.contains("enabled = false"));
-        assert!(rendered.contains("backend = "));
-        assert!(rendered.contains("ollama_url = "));
-    }
+    app::run(cli, config_path, config).await
 }
